% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_base.R
\name{apply_subset}
\alias{apply_subset}
\title{Apply subset to data}
\usage{
apply_subset(data, subset, data_type = c("matrix", "array"))
}
\arguments{
\item{data}{Data array or matrix}

\item{subset}{Subset specification from handle}

\item{data_type}{Type of data structure ("matrix" or "array")}
}
\value{
Subsetted data
}
\description{
Apply subset to data
}
\keyword{internal}
