% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_coercion.R
\name{as_dense_mat}
\alias{as_dense_mat}
\title{Coerce object to a dense matrix}
\usage{
as_dense_mat(obj)
}
\arguments{
\item{obj}{Object to coerce.}
}
\value{
A matrix.
}
\description{
Provides a simple S3 generic used by write adapters to obtain
a time-by-voxel matrix representation. Packages can implement
methods for their own classes.
}
\keyword{internal}
