% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl_verbs.R
\name{as_pipeline}
\alias{as_pipeline}
\title{Initiate an LNA pipeline}
\usage{
as_pipeline(x, run_ids = NULL, chunk_mb_suggestion = NULL)
}
\arguments{
\item{x}{Data object or list of run data.}

\item{run_ids}{Optional character vector of run identifiers.}

\item{chunk_mb_suggestion}{Optional numeric hint for chunk size.}
}
\value{
A configured `lna_pipeline` object.
}
\description{
Creates a new `lna_pipeline` object and sets its input using
`lna_pipeline$set_input()`.
}
