% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_coercion.R
\name{as_vol4d}
\alias{as_vol4d}
\title{Coerce object to a 4-D array}
\usage{
as_vol4d(obj)
}
\arguments{
\item{obj}{Object to coerce.}
}
\value{
A 4-D array.
}
\description{
Provides a simple S3 generic used by write adapters when
volumetric data is required.
}
\keyword{internal}
