% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sobel_rcpp_helpers.R
\name{benchmark_sobel3d_sizes}
\alias{benchmark_sobel3d_sizes}
\title{Benchmark Sobel3D on different volume sizes}
\usage{
benchmark_sobel3d_sizes()
}
\description{
Benchmark Sobel3D on different volume sizes
}
