% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrbf_helpers.R
\name{check_edge_adaptive_performance}
\alias{check_edge_adaptive_performance}
\title{Check edge-adaptive HRBF sampling availability and performance}
\usage{
check_edge_adaptive_performance(dims, warn_large = TRUE)
}
\arguments{
\item{dims}{Vector of 3D dimensions}

\item{warn_large}{Logical, whether to warn about large volumes}
}
\value{
List with recommendations and performance info
}
\description{
Helper to determine if edge-adaptive sampling should be used based on
volume size and available acceleration.
}
