% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_registry.R
\name{clear_transform_registry}
\alias{clear_transform_registry}
\title{Clear transform registry}
\usage{
clear_transform_registry()
}
\value{
Invisible NULL
}
\description{
Clear transform registry
}
\keyword{internal}
