\name{delta.default}
\alias{delta.default}
\title{Default method for delta DSL verb}
\usage{
delta.default(data_or_pipe, ...)
}
\description{
This is an alias for \code{\link{delta}} that provides the default method 
for the delta DSL verb. It is exported to support S3 method dispatch.
}
\seealso{\code{\link{delta}}}
\keyword{internal}