% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimental_api.R
\name{dequantize_data}
\alias{dequantize_data}
\title{Dequantize Data}
\usage{
dequantize_data(quant_result)
}
\arguments{
\item{quant_result}{Result from quantize_data()}
}
\value{
Reconstructed data (not yet implemented)
}
\description{
Reconstruct original data from quantized representation.
Note: This function is not yet implemented.
}
