% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_basis.R
\name{.dct_basis}
\alias{.dct_basis}
\title{Temporal Basis Generation Functions}
\usage{
.dct_basis(n_time, n_basis)
}
\description{
This file contains all functions related to generating temporal basis matrices
for different basis types (DCT, B-spline, DPSS, polynomial, wavelet).
Generate an orthonormal DCT basis matrix
}
\keyword{internal}
