% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_rle.R
\name{.encode_rle}
\alias{.encode_rle}
\title{Run-Length Encode a Vector}
\usage{
.encode_rle(vec)
}
\arguments{
\item{vec}{Vector to encode.}
}
\value{
A matrix with columns "lengths" and "values".
}
\description{
Converts a vector into a two-column matrix with columns
"lengths" and "values" using base `rle`.
}
\keyword{internal}
