% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_hdf5.R
\name{.is_valid_h5_object}
\alias{.is_valid_h5_object}
\title{HDF5 Attribute Read/Write Helpers}
\usage{
.is_valid_h5_object(h5_obj)
}
\description{
Provides internal functions for reading, writing, checking existence,
  and deleting attributes associated with HDF5 objects (groups or datasets).
}
\keyword{internal}
