% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{.LNA_PATHS}
\alias{.LNA_PATHS}
\title{LNA Constants}
\format{
An object of class \code{list} of length 9.
}
\usage{
.LNA_PATHS
}
\description{
Central location for package-wide constants, magic numbers, and default values.
These values are used throughout the package and can be referenced instead of
hardcoding values in multiple locations.
}
\keyword{internal}
