% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_quant.R
\name{.quantize_global}
\alias{.quantize_global}
\title{Compute quantization parameters globally}
\usage{
.quantize_global(x, bits, method, center)
}
\description{
Compute quantization parameters globally
}
\keyword{internal}
