% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_quant.R
\name{.quantize_voxel}
\alias{.quantize_voxel}
\title{Compute quantization parameters per voxel (time series)}
\usage{
.quantize_voxel(x, bits, method, center)
}
\description{
Compute quantization parameters per voxel (time series)
}
\keyword{internal}
