% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_quant.R
\name{.quantize_voxel_core}
\alias{.quantize_voxel_core}
\title{Quantize voxel-wise time series}
\usage{
.quantize_voxel_core(x, bits, method, center, collect_clip = FALSE)
}
\description{
Internal helper that performs quantization for each voxel. When
`collect_clip` is `TRUE`, the number of clipped samples is also returned.
}
\keyword{internal}
