\name{embed.default}
\alias{embed.default}
\title{Default method for embed DSL verb}
\usage{
embed.default(data_or_pipe, ...)
}
\description{
This is an alias for \code{\link{embed}} that provides the default method 
for the embed DSL verb. It is exported to support S3 method dispatch.
}
\seealso{\code{\link{embed}}}
\keyword{internal}