% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_core_read.R
\name{finalize_handle_for_read}
\alias{finalize_handle_for_read}
\title{Finalize handle for return}
\usage{
finalize_handle_for_read(handle, output_dtype, allow_plugins, file)
}
\description{
Ensures dtype support and stores metadata.
}
\keyword{internal}
