% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispatch.R
\name{forward_step.default}
\alias{forward_step.default}
\title{Default method for forward_step.}
\usage{
\method{forward_step}{default}(type, desc, handle)
}
\arguments{
\item{type}{(character) The transform type.}

\item{desc}{(list) The transform descriptor.}

\item{handle}{(DataHandle) The data handle.}
}
\value{
Throws an error because no specific method is defined.
}
\description{
Default method for forward_step.
}
\keyword{internal}
