% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_delta.R
\name{forward_step.delta}
\alias{forward_step.delta}
\title{Delta Transform - Forward Step}
\usage{
\method{forward_step}{delta}(type, desc, handle)
}
\description{
Computes first-order differences along a specified axis. When
\code{coding_method = "rle"} the differences are compressed using the
\code{rle} package and stored as a two column matrix of run lengths and
values.
}
\keyword{internal}
