% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_embed_transfer_hrbf_basis.R
\name{forward_step.embed.transfer_hrbf_basis}
\alias{forward_step.embed.transfer_hrbf_basis}
\title{Transfer HRBF Basis - Forward Step}
\usage{
\method{forward_step}{embed.transfer_hrbf_basis}(type, desc, handle)
}
\description{
Applies an empirical HRBF basis from a source LNA file to new data and
stores the resulting coefficient matrix.
}
\keyword{internal}
