% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_registry.R
\name{get_transform_capabilities}
\alias{get_transform_capabilities}
\title{Get transform capabilities}
\usage{
get_transform_capabilities(type)
}
\arguments{
\item{type}{Transform type}
}
\value{
Named list of capabilities
}
\description{
Get transform capabilities
}
