% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_registry.R
\name{get_transform_info}
\alias{get_transform_info}
\title{Get transform information}
\usage{
get_transform_info(type)
}
\arguments{
\item{type}{Transform type}
}
\value{
Transform info list or NULL if not found
}
\description{
Get transform information
}
