% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_hdf5.R
\name{guess_h5_type}
\alias{guess_h5_type}
\title{Guess an HDF5 datatype for an R object}
\usage{
guess_h5_type(x)
}
\arguments{
\item{x}{R object to inspect.}
}
\value{
An `H5T` datatype object.
}
\description{
Guess an HDF5 datatype for an R object
}
\keyword{internal}
