% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_hdf5.R
\name{h5_attr_delete}
\alias{h5_attr_delete}
\title{Delete an attribute from an HDF5 object.}
\usage{
h5_attr_delete(h5_obj, name)
}
\arguments{
\item{h5_obj}{An H5Group or H5D object from hdf5r.}

\item{name}{The name of the attribute to delete.}
}
\value{
Invisibly returns NULL.
}
\description{
Delete an attribute from an HDF5 object.
}
\details{
Does nothing if the attribute does not exist.
}
