% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_hdf5.R
\name{h5_attr_exists}
\alias{h5_attr_exists}
\title{Check if an attribute exists on an HDF5 object.}
\usage{
h5_attr_exists(h5_obj, name)
}
\arguments{
\item{h5_obj}{An H5Group or H5D object from hdf5r.}

\item{name}{The name of the attribute.}
}
\value{
Logical TRUE if the attribute exists, FALSE otherwise.
}
\description{
Check if an attribute exists on an HDF5 object.
}
