% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_hdf5.R
\name{h5_attr_read}
\alias{h5_attr_read}
\title{Read an attribute from an HDF5 object.}
\usage{
h5_attr_read(h5_obj, name)
}
\arguments{
\item{h5_obj}{An H5Group or H5D object from hdf5r.}

\item{name}{The name of the attribute.}
}
\value{
The value of the attribute.
}
\description{
Read an attribute from an HDF5 object.
}
\details{
Throws an error if the attribute does not exist.
}
