% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_hdf5.R
\name{h5_attr_write}
\alias{h5_attr_write}
\title{Write an attribute to an HDF5 object.}
\usage{
h5_attr_write(h5_obj, name, value)
}
\arguments{
\item{h5_obj}{An H5Group or H5D object from hdf5r.}

\item{name}{The name of the attribute.}

\item{value}{The value to write (basic R types and simple vectors supported).}
}
\value{
Invisibly returns NULL.
}
\description{
Write an attribute to an HDF5 object.
}
\details{
Overwrites the attribute if it already exists.
}
