% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_hdf5.R
\name{h5_read_subset}
\alias{h5_read_subset}
\title{Read a subset of a dataset from an HDF5 group}
\usage{
h5_read_subset(h5_group, path, index)
}
\arguments{
\item{h5_group}{An `H5Group` object used as the starting location for `path`.}

\item{path}{Character string giving the dataset path relative to `h5_group`.}

\item{index}{List of indices for each dimension as accepted by `hdf5r`.}
}
\value{
The selected subset of the dataset.
}
\description{
Read a subset of a dataset from an HDF5 group
}
\details{
Throws an error if the dataset does not exist or reading fails.
}
