\name{hrbf.default}
\alias{hrbf.default}
\title{Default method for hrbf DSL verb}
\usage{
hrbf.default(data_or_pipe, levels = NULL, ...)
}
\description{
This is an alias for \code{\link{hrbf}} that provides the default method 
for the hrbf DSL verb. It is exported to support S3 method dispatch.
}
\seealso{\code{\link{hrbf}}}
\keyword{internal}