% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_delta.R
\name{invert_step.delta}
\alias{invert_step.delta}
\title{Delta Transform - Inverse Step}
\usage{
\method{invert_step}{delta}(type, desc, handle)
}
\description{
Reconstructs data from delta representation stored in HDF5.
}
\keyword{internal}
