% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_embed_transfer_hrbf_basis.R
\name{invert_step.embed.transfer_hrbf_basis}
\alias{invert_step.embed.transfer_hrbf_basis}
\title{Transfer HRBF Basis - Inverse Step}
\usage{
\method{invert_step}{embed.transfer_hrbf_basis}(type, desc, handle)
}
\description{
Reconstructs dense data from coefficients using an HRBF basis stored in a
source LNA file.
}
\keyword{internal}
