% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispatch.R
\name{invert_step}
\alias{invert_step}
\title{Apply an inverse transform step.}
\usage{
invert_step(type, desc, handle)
}
\arguments{
\item{type}{(character) The type identifier of the transform (e.g., "mask", "pca").}

\item{desc}{(list) The parsed JSON descriptor for this transform step.}

\item{handle}{(DataHandle) The current data handle.}
}
\value{
An updated `DataHandle` object after applying the inverse step.
}
\description{
Apply an inverse transform step.
}
