% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_progress.R
\name{is_progress_globally_enabled}
\alias{is_progress_globally_enabled}
\title{Check if progressr handlers are active and not void}
\usage{
is_progress_globally_enabled()
}
\value{
Logical, TRUE if progress reporting is effectively enabled, FALSE otherwise.
}
\description{
This function checks if progressr has any active handlers and
ensures that not all of them are "void" handlers (which silence output).
}
\keyword{internal}
