% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_registry.R
\name{is_transform_registered}
\alias{is_transform_registered}
\title{Check if transform is registered}
\usage{
is_transform_registered(type)
}
\arguments{
\item{type}{Transform type}
}
\value{
Logical
}
\description{
Check if transform is registered
}
