% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_aggregate_runs.R
\name{lna_default.aggregate_runs}
\alias{lna_default.aggregate_runs}
\title{Default parameters for aggregate_runs}
\usage{
lna_default.aggregate_runs()
}
\description{
Default parameters for aggregate_runs
}
\keyword{internal}
