\name{lna_default.basis.empirical_hrbf_compressed}
\alias{lna_default.basis.empirical_hrbf_compressed}
\title{Default parameters for the 'basis.empirical_hrbf_compressed' transform}
\usage{
lna_default.basis.empirical_hrbf_compressed()
}
\description{
Convenience wrapper around \code{\link{default_params}}("basis.empirical_hrbf_compressed").
}
\seealso{\code{\link{default_params}}}
