% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_myorg_aggregate_runs.R
\name{lna_default.myorg.aggregate_runs}
\alias{lna_default.myorg.aggregate_runs}
\title{Default parameters for myorg.aggregate_runs}
\usage{
lna_default.myorg.aggregate_runs()
}
\description{
Default parameters for myorg.aggregate_runs
}
\keyword{internal}
