% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_sparsepca.R
\name{lna_default.myorg.sparsepca}
\alias{lna_default.myorg.sparsepca}
\title{Default parameters for myorg.sparsepca}
\usage{
lna_default.myorg.sparsepca()
}
\description{
Default parameters for myorg.sparsepca including a default seed for
reproducibility.
}
\keyword{internal}
