\name{lna_default.spat.haar_octwave}
\alias{lna_default.spat.haar_octwave}
\title{Default parameters for the 'spat.haar_octwave' transform}
\usage{
lna_default.spat.haar_octwave()
}
\description{
Convenience wrapper around \code{\link{default_params}}("spat.haar_octwave").
}
\seealso{\code{\link{default_params}}}
