\name{lna_default.spat.hrbf_project}
\alias{lna_default.spat.hrbf_project}
\title{Default parameters for the 'spat.hrbf_project' transform}
\usage{
lna_default.spat.hrbf_project()
}
\description{
Convenience wrapper around \code{\link{default_params}}("spat.hrbf_project").
}
\seealso{\code{\link{default_params}}}
