% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_error.R
\docType{data}
\name{lna_error_closed_reader}
\alias{lna_error_closed_reader}
\title{Error thrown when `lna_reader` methods are called after the reader is closed.}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
lna_error_closed_reader
}
\description{
Error thrown when `lna_reader` methods are called after the reader is closed.
}
\keyword{internal}
