% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl_templates.R
\docType{data}
\name{lna_template_registry_env}
\alias{lna_template_registry_env}
\title{DSL Template Registry}
\format{
An environment containing the template registry.
}
\usage{
lna_template_registry_env
}
\description{
Provides functions to register and apply pipeline templates.
}
\keyword{datasets}
