% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl_verbs.R
\name{lna_write}
\alias{lna_write}
\title{Execute an LNA pipeline}
\usage{
lna_write(pipeline_obj, file, ..., .verbose = TRUE, .checksum = "sha256")
}
\arguments{
\item{pipeline_obj}{An `lna_pipeline` object.}

\item{file}{Path to the output `.h5` file.}

\item{...}{Additional arguments forwarded to `write_lna()` such as
`header`, `mask`, or `plugins`.}

\item{.verbose}{Logical flag controlling verbosity (currently unused).}

\item{.checksum}{Checksum mode forwarded to `write_lna()`.}
}
\value{
The result returned by `write_lna()`.
}
\description{
Translates an `lna_pipeline` object into a call to `write_lna()` and
materialises the resulting archive on disk.
}
