% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facade.R
\name{LNAFacade}
\alias{LNAFacade}
\title{Simple facade for LNA read/write}
\description{An R6 class providing \code{write()} and \code{read()} methods that
wrap \code{write_lna()} and \code{read_lna()}. The last written file path is
stored in \code{last_output}.}
\section{Methods}{
\itemize{
  \item \code{new(transform_params = list())} create a new facade
  \item \code{write(x, file, transforms, ...)} write data via \code{write_lna}
  \item \code{read(file, ...)} read data via \code{read_lna}
}}
\examples{
f <- LNAFacade$new()
tmp <- tempfile(fileext = ".h5")
f$write(array(1, dim = c(1,1,1)), tmp, transforms = character())
f$read(tmp)
}
