% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_core_read.R
\name{normalize_allow_plugins}
\alias{normalize_allow_plugins}
\title{Normalize allow_plugins argument}
\usage{
normalize_allow_plugins(choice)
}
\description{
Ensures prompt mode degrades to installed when not interactive.
}
\keyword{internal}
