\name{pca.default}
\alias{pca.default}
\title{Default method for pca DSL verb}
\usage{
pca.default(data_or_pipe, ...)
}
\description{
This is an alias for \code{\link{pca}} that provides the default method 
for the pca DSL verb. It is exported to support S3 method dispatch.
}
\seealso{\code{\link{pca}}}
\keyword{internal}