% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_core_read.R
\name{process_run_core_read}
\alias{process_run_core_read}
\title{Process a single run for core_read}
\usage{
process_run_core_read(
  rid,
  h5,
  runs,
  subset_params,
  transforms,
  tf_group,
  validate,
  output_dtype,
  allow_plugins,
  file
)
}
\description{
Constructs a `DataHandle`, applies transforms and finalizes the result.
}
\keyword{internal}
