% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_json.R
\name{read_json_descriptor}
\alias{read_json_descriptor}
\title{JSON Read/Write Helpers for LNA Descriptors}
\usage{
read_json_descriptor(h5_group, name)
}
\arguments{
\item{h5_group}{An H5Group object from hdf5r.}

\item{name}{The name of the HDF5 dataset containing the JSON string.}
}
\value{
A list object parsed from the JSON string.
}
\description{
Provides internal functions for reading and writing transform
  descriptors stored as JSON strings within HDF5 datasets.
}
\details{
Assumes the dataset stores a single UTF-8 string (potentially variable length). Numeric
  values in the JSON are coerced with `as.numeric()` so that whole-number values are not
  returned as integers.
}
\keyword{HDF5}
\keyword{JSON}
\keyword{Read}
\keyword{a}
\keyword{an}
\keyword{descriptor}
\keyword{from}
\keyword{group.}
\keyword{internal}
