% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_registry.R
\name{register_transform}
\alias{register_transform}
\title{Register a transform}
\usage{
register_transform(type, metadata = list(), validate = TRUE)
}
\arguments{
\item{type}{Character transform type name}

\item{metadata}{List of metadata about the transform}

\item{validate}{Whether to validate the transform has required methods}
}
\value{
Invisible NULL
}
\description{
Register a transform
}
\keyword{internal}
