% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_base.R
\name{resolve_input_key}
\alias{resolve_input_key}
\title{Resolve input key for transform}
\usage{
resolve_input_key(handle, desc, preferred_keys = character())
}
\arguments{
\item{handle}{DataHandle object}

\item{desc}{Transform descriptor}

\item{preferred_keys}{Character vector of preferred keys to check first}
}
\value{
Character string of resolved input key
}
\description{
Resolve input key for transform
}
\keyword{internal}
