% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_core_read.R
\name{resolve_runs_for_read}
\alias{resolve_runs_for_read}
\title{Resolve runs for reading}
\usage{
resolve_runs_for_read(h5, run_id, lazy)
}
\description{
Determines which run identifiers to process and handles lazy mode.
}
\keyword{internal}
