\name{temporal.default}
\alias{temporal.default}
\title{Default method for temporal DSL verb}
\usage{
temporal.default(data_or_pipe, n_basis = NULL, kind = "bspline", ...)
}
\description{
This is an alias for \code{\link{temporal}} that provides the default method 
for the temporal DSL verb. It is exported to support S3 method dispatch.
}
\seealso{\code{\link{temporal}}}
\keyword{internal}