% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_temporal.R
\name{temporal_basis}
\alias{temporal_basis}
\title{Generate temporal basis matrix}
\usage{
temporal_basis(kind, n_time, n_basis, ...)
}
\arguments{
\item{kind}{Character scalar identifying the basis type.}

\item{n_time}{Integer number of time points.}

\item{n_basis}{Integer number of basis functions.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A basis matrix with dimensions \code{n_time x n_basis}.
}
\description{
Dispatches on \code{kind} to create a temporal basis. Package authors can
extend this generic by defining methods named \code{temporal_basis.<kind>}.
}
