% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_project.R
\name{temporal_project}
\alias{temporal_project}
\title{Temporal Projection Functions}
\usage{
temporal_project(kind, basis, X, ...)
}
\description{
This file contains all temporal projection functions that project data
onto temporal bases with optional thresholding/denoising.
Project data onto a temporal basis
}
\details{
Each temporal basis kind can implement customised projection logic.
The default assumes an orthonormal basis and uses `crossprod`.
}
\keyword{internal}
