% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_meta.R
\name{transform_min_dims}
\alias{transform_min_dims}
\alias{transform_min_dims.default}
\alias{transform_min_dims.quant}
\title{Minimum input dimensionality for a transform}
\usage{
transform_min_dims(type)
}
\arguments{
\item{type}{Character transform type.}
}
\description{
Returns the minimum number of dimensions required by the given transform's
forward step. The default is 3.
}
\keyword{internal}
