% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_registry.R
\name{transform_summary}
\alias{transform_summary}
\title{Get transform summary}
\usage{
transform_summary()
}
\value{
Data frame summarizing registered transforms
}
\description{
Get transform summary
}
