% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_base.R
\name{validators}
\alias{validators}
\alias{validate_integer_range}
\alias{validate_enum}
\alias{validate_probability}
\title{Create validator functions for common patterns}
\usage{
validate_integer_range(min, max)

validate_enum(allowed_values)

validate_probability(value, param_name)
}
\description{
Create validator functions for common patterns
}
\keyword{internal}
