% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_json.R
\name{write_json_descriptor}
\alias{write_json_descriptor}
\title{Write a JSON descriptor to an HDF5 group.}
\usage{
write_json_descriptor(h5_group, name, desc_list)
}
\arguments{
\item{h5_group}{An H5Group object from hdf5r.}

\item{name}{The name of the HDF5 dataset to create or overwrite.}

\item{desc_list}{A list object to be converted to JSON.}
}
\value{
Invisibly returns NULL.
}
\description{
Write a JSON descriptor to an HDF5 group.
}
\details{
Writes the list as a JSON string to a scalar HDF5 dataset with
  a variable-length string datatype (UTF-8). Overwrites existing dataset
  with the same name.
}
