% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_temporal.R
\name{.polynomial_basis}
\alias{.polynomial_basis}
\title{Generate orthogonal polynomial basis matrix}
\usage{
.polynomial_basis(n_time, n_basis)
}
\description{
Generate orthogonal polynomial basis matrix
}
\keyword{internal}
