% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_delta.R
\name{forward_step.delta}
\alias{forward_step.delta}
\title{Delta Transform - Forward Step}
\usage{
\method{forward_step}{delta}(type, desc, handle)
}
\description{
Computes first-order differences along a specified axis and optionally
run-length encodes the result.
}
\keyword{internal}
