% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_embed.R
\name{forward_step.embed}
\alias{forward_step.embed}
\title{Embed Transform - Forward Step}
\usage{
\method{forward_step}{embed}(type, desc, handle)
}
\description{
Projects data onto a pre-computed basis matrix.
}
\keyword{internal}
