% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_quant.R
\name{forward_step.quant}
\alias{forward_step.quant}
\title{Quantization Transform - Forward Step}
\usage{
\method{forward_step}{quant}(type, desc, handle)
}
\description{
Applies simple min-max quantization to numeric data.
}
\keyword{internal}
