% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_spat_hrbf_project.R
\name{forward_step.spat.hrbf_project}
\alias{forward_step.spat.hrbf_project}
\title{HRBF Project Transform - Forward Step}
\usage{
\method{forward_step}{spat.hrbf_project}(type, desc, handle)
}
\description{
Projects input data onto an analytic HRBF basis without storing the basis
or centres in the file. Only the coefficient matrix is returned in the stash.
}
\keyword{internal}
