% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_embed.R
\name{invert_step.embed}
\alias{invert_step.embed}
\title{Embed Transform - Inverse Step}
\usage{
\method{invert_step}{embed}(type, desc, handle)
}
\description{
Reconstructs data from embedding coefficients using a stored basis matrix.
}
\keyword{internal}
