% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_sparsepca.R
\name{invert_step.myorg.sparsepca}
\alias{invert_step.myorg.sparsepca}
\title{Sparse PCA Transform - Inverse Step}
\usage{
\method{invert_step}{myorg.sparsepca}(type, desc, handle)
}
\arguments{
\item{storage_order}{Character string specifying the orientation of the
basis matrix. Either "component_x_voxel" (default) or
"voxel_x_component".}
}
\description{
Reconstructs data from the sparse PCA coefficients and basis matrix.
}
\keyword{internal}
