% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_sparsepca.R
\name{invert_step.myorg.sparsepca}
\alias{invert_step.myorg.sparsepca}
\title{Sparse PCA Transform - Inverse Step}
\usage{
\method{invert_step}{myorg.sparsepca}(type, desc, handle)
}
\description{
Reconstructs data from the sparse PCA coefficients and basis matrix.
}
\keyword{internal}
