% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_quant.R
\name{invert_step.quant}
\alias{invert_step.quant}
\title{Quantization Transform - Inverse Step}
\usage{
\method{invert_step}{quant}(type, desc, handle)
}
\description{
Reconstructs data from quantized representation stored in HDF5.
}
\keyword{internal}
