% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_quant.R
\name{invert_step.quant}
\alias{invert_step.quant}
\title{Quantization Transform - Inverse Step}
\usage{
\method{invert_step}{quant}(type, desc, handle)
}
\description{
Reads quantised values, validates the \code{quant_bits} attribute and
reconstructs the original data using stored scale and offset.
}
\keyword{internal}
