% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_spat_hrbf_project.R
\name{invert_step.spat.hrbf_project}
\alias{invert_step.spat.hrbf_project}
\title{HRBF Project Transform - Inverse Step}
\usage{
\method{invert_step}{spat.hrbf_project}(type, desc, handle)
}
\description{
Reconstructs data from HRBF coefficients by regenerating the analytic
basis on-the-fly or loading stored centres if provided.
}
\keyword{internal}
