% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_reports.R
\name{lna_get_quant_report}
\alias{lna_get_quant_report}
\title{Retrieve a quantization report}
\usage{
lna_get_quant_report(lna_file, transform_index_or_name)
}
\arguments{
\item{lna_file}{Path to an LNA file.}
\item{transform_index_or_name}{Integer index or descriptor name identifying the transform.}
}
\value{
A list parsed from the quantization report.
}
\description{
Convenience wrapper around \code{lna_get_transform_report()}.
}
\examples
{\donttest{
tmp <- tempfile(fileext = ".h5")
arr <- array(runif(6), dim = c(2,3))
write_lna(arr, tmp, transforms = "quant")
rep <- lna_get_quant_report(tmp, 0)
}}
