% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl_registry.R
\docType{data}
\name{lna_verb_registry_env}
\alias{lna_verb_registry_env}
\title{DSL Verb Registry}
\format{
An environment containing the verb registry.
}
\usage{
lna_verb_registry_env
}
\description{
Provides a registry for DSL verbs so that external packages can
expose additional pipeline verbs.  The registry simply maps a verb
function name to the corresponding LNA transform type.
}
\keyword{datasets}
