% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_temporal.R
\name{temporal_project}
\alias{temporal_project}
\title{Project data onto a temporal basis}
\usage{
temporal_project(kind, basis, X, ...)
}
\description{
Each temporal basis kind can implement customised projection logic.
The default assumes an orthonormal basis and uses `crossprod`.
}
\keyword{internal}
