% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_reconstruct.R
\name{temporal_reconstruct}
\alias{temporal_reconstruct}
\title{Temporal Reconstruction Functions}
\usage{
temporal_reconstruct(kind, basis, coeff, ...)
}
\description{
This file contains all temporal reconstruction functions that reconstruct
data from temporal coefficients using various basis types.
Reconstruct data from temporal coefficients
}
\details{
Mirrors `temporal_project` for the inverse operation. The default
simply multiplies the basis by the coefficient matrix.
}
\keyword{internal}
