% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_meta.R
\name{transform_min_dims}
\alias{transform_min_dims}
\title{Minimum input dimensionality for a transform}
\usage{
transform_min_dims(type)
}
\arguments{
\item{type}{Character transform type.}
}
\description{
Provides the minimum number of dimensions required by a transform's
forward step. Packages can define methods for their own transforms.
The default requirement is 3 dimensions if a specific transform is not listed.
}
