% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_write.R
\name{validate_input_data}
\alias{validate_input_data}
\title{Validate input data}
\usage{
validate_input_data(x)
}
\arguments{
\item{x}{Input object for `core_write`.}
}
\description{
Ensures that `x` (or each element of a list `x`) has at least three
dimensions. This prevents ambiguous handling of 2D matrices.
}
\keyword{internal}
